gguf_rs_lib::tensor::data

Enum TensorData

pub enum TensorData {
    Owned(Vec<u8>),
    Borrowed(&'static [u8]),
    Shared(Arc<Vec<u8>>),
    Mapped {
        mmap: Arc<Mmap>,
        offset: usize,
        length: usize,
    },
    Empty,
}

Container for tensor data with different storage backends

Variants

Owned(Vec<u8>)

Owned byte vector (most common case)

Borrowed(&'static [u8])

Reference to borrowed data (zero-copy when possible)

Shared(Arc<Vec<u8>>)

Shared reference-counted data

Mapped

Memory-mapped data (when using mmap feature)

Fields

mmap: Arc<Mmap>

The memory map

offset: usize

Offset within the map

length: usize

Length of the data

Empty

Empty/uninitialized data

Implementations

impl TensorData

pub fn new_owned(data: Vec<u8>) -> Self

Create new owned tensor data

pub fn new_borrowed(data: &'static [u8]) -> Self

Create new borrowed tensor data

pub fn new_shared(data: Vec<u8>) -> Self

Create new shared tensor data

pub fn empty() -> Self

Create empty tensor data

pub fn zeros(size: usize) -> Self

Create tensor data with specified size filled with zeros

pub fn filled(size: usize, value: u8) -> Self

Create tensor data with specified size filled with a value

pub fn new_mapped(mmap: Arc<Mmap>, offset: usize, length: usize) -> Result<Self>

Create memory-mapped tensor data

pub fn len(&self) -> usize

Get the length of the tensor data in bytes

pub fn is_empty(&self) -> bool

Check if the tensor data is empty

pub fn as_slice(&self) -> &[u8]

Get a slice of the tensor data

pub fn as_mut_slice(&mut self) -> Option<&mut [u8]>

Get mutable access to the data (only for owned data)

pub fn to_owned(&self) -> Vec<u8>

Convert to owned data (cloning if necessary)

pub fn into_owned(self) -> Vec<u8>

Take ownership of the data if possible, otherwise clone

pub fn storage_type(&self) -> TensorStorageType

Get the storage type

pub fn storage_info(&self) -> TensorDataInfo

Get storage information

pub fn is_aligned(&self) -> bool

Check if the data is properly aligned

pub fn is_aligned_to(&self, alignment: usize) -> bool

Check if the data is aligned to a specific boundary

pub fn alignment(&self) -> Option<usize>

Get the alignment of the data pointer

pub fn slice(&self, start: usize, length: usize) -> Result<TensorData>

Create a slice of the tensor data

pub fn concat(&self, other: &TensorData) -> TensorData

Concatenate with another tensor data

pub fn content_equals(&self, other: &TensorData) -> bool

Check if two tensor data instances have the same content

pub fn hex_preview(&self, max_bytes: usize) -> String

Get a hexadecimal representation of the first few bytes (for debugging)

pub fn validate(&self) -> Result<()>

Validate the tensor data for basic consistency

pub fn checksum(&self) -> u32

Calculate a simple checksum of the data (for integrity checking)

pub fn memory_usage(&self) -> TensorMemoryUsage

Get memory usage information