gguf_rs_lib::tensor::quantization

Struct QuantizationUtils

pub struct QuantizationUtils;

Utilities for working with quantization

Implementations

impl QuantizationUtils

pub fn all_quantized_types() -> Vec<TensorType>

Get all supported quantization types

pub fn recommend_quantization( model_size_gb: f32, target_quality: f32, ) -> TensorType

Get recommended quantization for model size and quality requirements

pub fn compare_formats(type_a: TensorType, type_b: TensorType) -> Ordering

Compare two quantization formats

pub fn find_closest_quantization(target_bits: f32) -> TensorType

Get the most similar quantization to a target bit rate

pub fn is_modern_quantization(tensor_type: TensorType) -> bool

Check if a quantization type is considered modern/recommended

pub fn get_quantization_family(tensor_type: TensorType) -> &'static str

Get quantization family (legacy, k-quant, i-quant)